+++ dojo/models/__init__.py
@@ -1749,24 +1749,7 @@ class Development_Environment(models.Model):
                  "url": reverse("edit_dev_env", args=(self.id,))}]
 
 
-class Sonarqube_Issue(models.Model):
-    key = models.CharField(max_length=30, unique=True, help_text=_("SonarQube issue key"))
-    status = models.CharField(max_length=20, help_text=_("SonarQube issue status"))
-    type = models.CharField(max_length=20, help_text=_("SonarQube issue type"))
-
-    def __str__(self):
-        return self.key
-
-
-class Sonarqube_Issue_Transition(models.Model):
-    sonarqube_issue = models.ForeignKey(Sonarqube_Issue, on_delete=models.CASCADE, db_index=True)
-    created = models.DateTimeField(auto_now_add=True, null=False)
-    finding_status = models.CharField(max_length=100)
-    sonarqube_status = models.CharField(max_length=50)
-    transitions = models.CharField(max_length=100)
-
-    class Meta:
-        ordering = ('-created', )
+from .sonarqube import Sonarqube_Issue, Sonarqube_Issue_Transition  # noqa: F401
 
 
 class Test(models.Model):
